# Use hash values to speed-up to comparing entities