--- a/db/schema.sql
+++ b/db/schema.sql
@@ -38,6 +38,11 @@ create table if not exists tuples (
 	_id   text    not null,
 	_rev  integer not null,
 
+	-- Hash values of entities
+	--
+	_hash_l bigint,
+	_hash_r bigint,
+
 	-- Self references for computed tuples
 	--
 	_rid_l  text,
@@ -71,3 +76,6 @@ create table if not exists tuples (
 	constraint "tuples.check-r_entity_id" check (r_entity_id <> ''),
 	constraint "tuples.check-attrs" check (jsonb_typeof(attrs) = 'object')
 );
+
+create index "tuples.idx-ltr" on tuples using btree (space_id, _hash_l, relation, _hash_r, _id);
+create index "tuples.idx-rtl" on tuples using btree (space_id, _hash_r, relation, strand, _hash_l, _id);