--- a/db/schema.sql
+++ b/db/schema.sql
@@ -77,5 +77,4 @@ create table if not exists tuples (
 	constraint "tuples.check-attrs" check (jsonb_typeof(attrs) = 'object')
 );
 
-create index "tuples.idx-ltr" on tuples using btree (space_id, _l_hash, relation, _r_hash, _id);
-create index "tuples.idx-rtl" on tuples using btree (space_id, _r_hash, relation, strand, _l_hash, _id);
+create index "tuples.idx-rtl" on tuples using btree (space_id, _r_hash, relation, _l_hash, strand, _id);